feat(article): move article examples to design folder and create new entrypoint for examples
