+++ apps/design-land/tsconfig.app.json
@@ -3,7 +3,8 @@
   "compilerOptions": {
     "outDir": "../../out-tsc/app",
     "paths": {
-      "@daffodil/design": ["libs/design/src"]
+      "@daffodil/design": ["libs/design/src"],
+      "@daffodil/design/article/examples": ["libs/design/article/examples/src"]
     },
     "types": []
   },
