+++ libs/design/article/examples/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "@daffodil/design/article/examples"
+}
