+++ libs/design/article/examples/src/article-blockquote/article-blockquote.component.html
