+++ libs/design/article/examples/src/article-blockquote/article-blockquote.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'article-blockquote',
