+++ libs/design/article/examples/src/article-code-block/article-code-block.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'article-code-block',
