+++ libs/design/article/examples/src/article-code-inline/article-code-inline.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'article-code-inline',
