+++ libs/design/article/examples/src/article-headings/article-headings.component.ts
