+++ libs/design/article/examples/src/article-hr/article-hr.component.html
