+++ libs/design/article/examples/src/article-hr/article-hr.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'article-hr',
