+++ libs/design/article/examples/src/article-lead/article-lead.component.html
