+++ libs/design/article/examples/src/article-meta/article-meta.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'article-meta',
